+++ src/main/java/rbasamoyai/createbigcannons/cannon_control/ControlPitchContraption.java
@@ -34,6 +34,7 @@ public interface ControlPitchContraption {
 
 	interface Block extends ControlPitchContraption {
 		BlockPos getControllerBlockPos();
+        void markForReassembly();
 	}
 
 }