+++ src/main/java/rbasamoyai/createbigcannons/mixin/CBCMixinPlugin.java
@@ -1,0 +1,41 @@
+package rbasamoyai.createbigcannons.mixin;
+
+import java.util.List;
+import java.util.Set;
+
+import org.objectweb.asm.tree.ClassNode;
+import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
+import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
+
+import rbasamoyai.createbigcannons.CBCModsNeoForge;
+
+public class CBCMixinPlugin implements IMixinConfigPlugin {
+
+    @Override
+    public void onLoad(String mixinPackage) {
+    }
+
+    @Override public String getRefMapperConfig() { return null; }
+
+    @Override
+    public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
+        if (mixinClassName.startsWith("rbasamoyai.createbigcannons.mixin.compat.simulated") && !CBCModsNeoForge.SIMULATED.isLoaded())
+            return false;
+        return true;
+    }
+
+    @Override
+    public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {
+    }
+
+    @Override public List<String> getMixins() { return null; }
+
+    @Override
+    public void preApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
+    }
+
+    @Override
+    public void postApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
+    }
+
+}